+++ app/src/main/java/com/opensource13/pillsogood/MainActivity.java
@@ -2,9 +2,11 @@ package com.opensource13.pillsogood;
 
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.graphics.Color;
 import android.os.Bundle;
+import android.provider.Browser;
 import android.text.format.Time;
 import android.util.AttributeSet;
 import android.util.Log;
@@ -21,15 +23,15 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatTextView;
 import androidx.recyclerview.widget.RecyclerView;
 
+//import com.opensource13.pillsogood.SQLite.DataAdapter;
+import com.opensource13.pillsogood.SQLite.SQLiteHelper;
+
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
 
 public class MainActivity extends AppCompatActivity {
-
-    //SQLiteDatabase database;
-    //String dbName =
 
     GridView monthView;
 
@@ -41,517 +43,116 @@ public class MainActivity extends AppCompatActivity {
 
     int curMonth;
 
+    String result;
+
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        // 월별 캘린더 뷰 객체 참조
-        monthView = (GridView) findViewById(R.id.monthView);
-        monthViewAdapter = new MonthAdapter(getApplicationContext());
-        monthView.setAdapter(monthViewAdapter);
+        SQLiteHelper helper;
+        SQLiteDatabase db;
+        helper = new SQLiteHelper(MainActivity.this);
+        db = helper.getWritableDatabase();
+        helper.onCreate(db);
+
+        //////////////////////////////////////////////////////////////////////////////////////////////////////////
+        //db.execSQL("insert into MYDRUG values('" + "감기약" + "','" + "수" + "','" + "12:00" + "')");
+        //db.execSQL("insert into MYDRUG values('" + "항생제" + "','" + "월" + "','" + "13:00" + "')"); //db insert code
+       // Cursor cursor = db.rawQuery("SELECT * FROM MYDRUG", null);
+       // result = "";
+       // while (cursor.moveToNext()) {
+            //String temp = cursor.getString(0);
+            //titleView.setText(cursor.getString(0));
+
+            //   Log.d("MainActivity", "drug_name : " + temp);
+          //   }
+
+            // 월별 캘린더 뷰 객체 참조
+            monthView = (GridView) findViewById(R.id.monthView);
+            monthViewAdapter = new MonthAdapter(getApplicationContext());
+            monthView.setAdapter(monthViewAdapter);
+
+            // 리스너 설정
+            monthView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+                @Override
+                public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
+                    // 현재 선택한 일자 정보 표시
+                    MonthItem curItem = (MonthItem) monthViewAdapter.getItem(position);
+                    int day = curItem.getDay();
+
+                    Log.d("MainActivity", "Selected : " + day);
 
-        // 리스너 설정
-        monthView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
-                // 현재 선택한 일자 정보 표시
-                MonthItem curItem = (MonthItem) monthViewAdapter.getItem(position);
-                int day = curItem.getDay();
-
-                Log.d("MainActivity", "Selected : " + day);
-            }
-        });
-
-        // 날짜 누르면 DaydruglistActivity로 넘어가는 이벤트 처리
-        monthView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
-                Intent intent = new Intent(MainActivity.this, DaydruglistActivity.class);
-                // intent.putExtra("image", logos[position]); // put image data in Intent
-                startActivity(intent); // start Intent
-            }
-        });
-
-        monthText = (TextView) findViewById(R.id.monthText);
-        setMonthText();
-
-        // 이전 월로 넘어가는 이벤트 처리
-        Button monthPrevious = (Button) findViewById(R.id.monthPrevious);
-        monthPrevious.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                monthViewAdapter.setPreviousMonth();
-                monthViewAdapter.notifyDataSetChanged();
-
-                setMonthText();
-            }
-        });
-
-        // 다음 월로 넘어가는 이벤트 처리
-        Button monthNext = (Button) findViewById(R.id.monthNext);
-        monthNext.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                monthViewAdapter.setNextMonth();
-                monthViewAdapter.notifyDataSetChanged();
-
-                setMonthText();
-            }
-        });
-
-        // "내 약" 버튼 누르면 MydrugActivity로 넘어가는 이벤트 처리
-        Button myPeel = (Button) findViewById(R.id.myPill);
-        myPeel.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), MydrugActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        /*
-        //오늘 날짜 세팅
-        long now = System.currentTimeMillis();
-        final Date date = new Date(now);
-        //연, 월, 일 따로 저장
-        final SimpleDateFormat curYearFormat = new SimpleDateFormat("yyyy", Locale.KOREA);
-        final SimpleDateFormat curMonthFormat = new SimpleDateFormat("MM", Locale.KOREA);
-        final SimpleDateFormat curDayFormat = new SimpleDateFormat("dd", Locale.KOREA);
-
-        //현재 날짜 텍스트뷰에 뿌리기
-        tvDate.setText(curYearFormat.format(date) + "/" + curMonthFormat.format(date));
-        */
-    }
-
-     //월 표시 텍스트 설정
-    private void setMonthText() {
-        curYear = monthViewAdapter.getCurYear();
-        curMonth = monthViewAdapter.getCurMonth();
-
-        monthText.setText(curYear + "년 " + (curMonth + 1) + "월");
-    }
-
-}
-
-//어댑터 객체 정의
-class MonthAdapter extends BaseAdapter {
-
-    public static final String TAG = "MonthAdapter";
-
-    Context mContext;
-
-    //public static int oddColor = Color.rgb(225, 225, 225);
-    //public static int headColor = Color.rgb(12, 32, 158);
-
-    private int selectedPosition = -1;
-
-    private MonthItem[] items;
-
-    private int countColumn = 7;
-
-    int mStartDay;
-    int startDay;
-    int curYear;
-    int curMonth;
-
-    int firstDay;
-    int lastDay;
-
-    Calendar mCalendar;
-    Calendar mCals;
-
-    boolean recreateItems = false;
-
-    public MonthAdapter(Context context) {
-        super();
-
-        mContext = context;
-
-        init();
-    }
-
-    public MonthAdapter(Context context, AttributeSet attrs) {
-        super();
-
-        mContext = context;
-
-        init();
-    }
-
-    private void init() {
-        items = new MonthItem[7 * 6];
-        mCalendar = Calendar.getInstance();
-        recalculate();
-        resetDayNumbers();
-    }
-
-
-    public void recalculate() {
-
-        // set to the first day of the month
-        mCalendar.set(Calendar.DAY_OF_MONTH, 1);
-
-        // get week day
-        int dayOfWeek = mCalendar.get(Calendar.DAY_OF_WEEK);
-        firstDay = getFirstDay(dayOfWeek);
-        Log.d(TAG, "firstDay : " + firstDay);
-
-        mStartDay = mCalendar.getFirstDayOfWeek();
-        curYear = mCalendar.get(Calendar.YEAR);
-        curMonth = mCalendar.get(Calendar.MONTH);
-        lastDay = getMonthLastDay(curYear, curMonth);
-
-        Log.d(TAG, "curYear : " + curYear + ", curMonth : " + curMonth + ", lastDay : " + lastDay);
-
-        int diff = mStartDay - Calendar.SUNDAY - 1;
-        startDay = getFirstDayOfWeek();
-        Log.d(TAG, "mStartDay : " + mStartDay + ", startDay : " + startDay);
-
-    }
-
-    public void setPreviousMonth() {
-        mCalendar.add(Calendar.MONTH, -1);
-        recalculate();
-
-        resetDayNumbers();
-        selectedPosition = -1;
-    }
-
-    public void setNextMonth() {
-        mCalendar.add(Calendar.MONTH, 1);
-        recalculate();
-
-        resetDayNumbers();
-        selectedPosition = -1;
-    }
-
-    public void resetDayNumbers() {
-        for (int i = 0; i < 42; i++) {
-            // calculate day number
-            int dayNumber = (i+1) - firstDay;
-            if (dayNumber < 1 || dayNumber > lastDay) {
-                dayNumber = 0;
-            }
-
-            // save as a data item
-            items[i] = new MonthItem(dayNumber);
-        }
-    }
-
-    private int getFirstDay(int dayOfWeek) {
-        int result = 0;
-        if (dayOfWeek == Calendar.SUNDAY) {
-            result = 0;
-        } else if (dayOfWeek == Calendar.MONDAY) {
-            result = 1;
-        } else if (dayOfWeek == Calendar.TUESDAY) {
-            result = 2;
-        } else if (dayOfWeek == Calendar.WEDNESDAY) {
-            result = 3;
-        } else if (dayOfWeek == Calendar.THURSDAY) {
-            result = 4;
-        } else if (dayOfWeek == Calendar.FRIDAY) {
-            result = 5;
-        } else if (dayOfWeek == Calendar.SATURDAY) {
-            result = 6;
-        }
-
-        return result;
-    }
-
-
-    public int getCurYear() {
-        return curYear;
-    }
-
-    public int getCurMonth() {
-        return curMonth;
-    }
-
-
-    public int getNumColumns() {
-        return 7;
-    }
-
-    public int getCount() {
-        return 7 * 6;
-    }
-
-    public Object getItem(int position) {
-        return items[position];
-    }
-
-    public long getItemId(int position) {
-        return position;
-    }
-
-    public View getView(int position, View convertView, ViewGroup parent) {
-        Log.d(TAG, "getView(" + position + ") called.");
-
-        MonthItemView itemView;
-       // ListItemView litemView;
-        //RecyclerView.ViewHolder holder = null;
-
-        if (convertView == null) {
-            itemView = new MonthItemView(mContext);
-        } else {
-            itemView = (MonthItemView) convertView;
-        }
-/*
-        if (convertView == null) {
-            litemView = new MonthItemView(mContext);
-        } else {
-            litemView = (MonthItemView) convertView;
-        }
-        */
-
-
-        // create a params
-        GridView.LayoutParams params = new GridView.LayoutParams(
-                GridView.LayoutParams.MATCH_PARENT, 196);
-
-        // calculate row and column
-        int rowIndex = position / countColumn;
-        int columnIndex = position % countColumn;
-
-        Log.d(TAG, "Index : " + rowIndex + ", " + columnIndex);
-
-        // set item data and properties
-        itemView.setItem(items[position]);
-        itemView.setLayoutParams(params);
-        itemView.setPadding(2, 2, 2, 2);
-
-        // set properties
-        itemView.setGravity(Gravity.LEFT);
-
-        if (columnIndex == 0) {
-            itemView.setTextColor(Color.RED);
-        }
-        else if (columnIndex == 6) {
-            itemView.setTextColor(Color.BLUE);
-        }
-        else {
-            itemView.setTextColor(Color.BLACK);
-        }
-
-        // set click background color
-        if (position == getSelectedPosition()) {
-            //Log.d("ClickBackground", "Selected : " );
-            //itemView.setTextColor(Color.RED);
-            itemView.setBackgroundColor(Color.YELLOW);
-        } else {
-            itemView.setBackgroundColor(Color.rgb(250, 250, 250));
-        }
-
-        //set today background color
-        mCals = Calendar.getInstance();
-        Integer today = mCals.get(Calendar.DAY_OF_MONTH);
-        String sToday = String.valueOf(today);
-
-        if (sToday.equals(getItem(position))) {
-            itemView.setBackgroundColor(Color.BLUE);
-            //itemView.setTextColor(Color.GREEN);
-        }
-
-        return itemView;
-    }
-
-
-    /**
-     * Get first day of week as android.text.format.Time constant.
-     * @return the first day of week in android.text.format.Time
-     */
-    public static int getFirstDayOfWeek() {
-        int startDay = Calendar.getInstance().getFirstDayOfWeek();
-        if (startDay == Calendar.SATURDAY) {
-            return Time.SATURDAY;
-        } else if (startDay == Calendar.MONDAY) {
-            return Time.MONDAY;
-        } else {
-            return Time.SUNDAY;
-        }
-    }
-
-
-    /**
-     * get day count for each month
-     *
-     * @param year
-     * @param month
-     * @return
-     */
-    private int getMonthLastDay(int year, int month){
-        switch (month) {
-            case 0:
-            case 2:
-            case 4:
-            case 6:
-            case 7:
-            case 9:
-            case 11:
-                return (31);
-
-            case 3:
-            case 5:
-            case 8:
-            case 10:
-                return (30);
-
-            default:
-                if(((year%4==0)&&(year%100!=0)) || (year%400==0) ) {
-                    return (29);   // 2월 윤년계산
-                } else {
-                    return (28);
                 }
-        }
-    }
+            });
 
-    /**
-     * set selected row
-     */
-    public void setSelectedPosition(int selectedPosition) {
-        this.selectedPosition = selectedPosition;
-    }
+            // 날짜 누르면 DaydruglistActivity로 넘어가는 이벤트 처리
+            monthView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+                @Override
+                public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
+                    Intent intent = new Intent(MainActivity.this, DaydruglistActivity.class);
+                    // intent.putExtra("image", logos[position]); // put image data in Intent
+                    startActivity(intent); // start Intent
+                }
+            });
 
-    /**
-     * get selected row
-     *
-     * @return
-     */
-    public int getSelectedPosition() {
-        return selectedPosition;
-    }
+            monthText = (TextView) findViewById(R.id.monthText);
+            setMonthText();
 
-}
+            // 이전 월로 넘어가는 이벤트 처리
+            Button monthPrevious = (Button) findViewById(R.id.monthPrevious);
+            monthPrevious.setOnClickListener(new View.OnClickListener() {
+                public void onClick(View v) {
+                    monthViewAdapter.setPreviousMonth();
+                    monthViewAdapter.notifyDataSetChanged();
 
-/**
- * 일자 정보를 담기 위한 클래스 정의
- */
-class MonthItem {
-    private int dayValue;
+                    setMonthText();
+                }
+            });
 
-    public MonthItem() {
+            // 다음 월로 넘어가는 이벤트 처리
+            Button monthNext = (Button) findViewById(R.id.monthNext);
+            monthNext.setOnClickListener(new View.OnClickListener() {
+                public void onClick(View v) {
+                    monthViewAdapter.setNextMonth();
+                    monthViewAdapter.notifyDataSetChanged();
+                    setMonthText();
+                }
+            });
 
-    }
-
-    public MonthItem(int day) {
-        dayValue = day;
-    }
-
-    public int getDay() {
-        return dayValue;
-    }
-
-    public void setDay(int day) {
-        this.dayValue = day;
-    }
-
-}
-
-/**
- * 약 리스트를 담기 위한 클래스 정의
-
-class ListItem {
-    private int peelValue;
-
-    public ListItem() {
-
-    }
-
-    public ListItem(int day) {
-        peelValue = day;
-    }
-
-    public int getDay() {
-        return peelValue;
-    }
-
-    public void setDay(int day) {
-        this.peelValue = day;
-    }
-
-}
- */
-
-/**
- * 일자에 표시하는 텍스트뷰 정의
- */
-class MonthItemView extends AppCompatTextView {
-
-    private MonthItem item;
-
-    public MonthItemView(Context context) {
-        super(context);
-
-        init();
-    }
-
-    public MonthItemView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-
-        init();
-    }
-
-    private void init() {
-
-        setBackgroundColor(Color.rgb(250, 250, 250));
-    }
-
-    public MonthItem getItem() {
-        return item;
-    }
-
-    public void setItem(MonthItem item) {
-        this.item = item;
-
-        int day = item.getDay();
-        if (day != 0) {
-            setText(String.valueOf(day));
-        } else {
-            setText("");
+            // "내 약" 버튼 누르면 MydrugActivity로 넘어가는 이벤트 처리
+            Button myPeel = (Button) findViewById(R.id.myPill);
+            myPeel.setOnClickListener(new View.OnClickListener() {
+                public void onClick(View v) {
+                    Intent intent = new Intent(getApplicationContext(), MydrugActivity.class);
+                    startActivity(intent);
+                }
+            });
         }
 
-    }
-}
+        //월 표시 텍스트 설정
+        private void setMonthText() {
+            curYear = monthViewAdapter.getCurYear();
+            curMonth = monthViewAdapter.getCurMonth();
 
-/**
- * 일자에 표시하는 텍스트뷰 정의
-
-class ListItemView extends AppCompatTextView {
-
-    private ListItem item;
-
-    public ListItemView(Context context) {
-        super(context);
-
-        init();
-    }
-
-    public ListItemView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-
-        init();
-    }
-
-    private void init() {
-
-        setBackgroundColor(Color.rgb(250, 250, 250));
-    }
-
-    public MonthItem getItem() {
-        return item;
-    }
-
-    public void setItem(MonthItem item) {
-        this.item = item;
-
-        int day = item.getDay();
-        if (day != 0) {
-            setText(String.valueOf(day));
-        } else {
-            setText("");
+            monthText.setText(curYear + "년 " + (curMonth + 1) + "월");
         }
 
-    }
+         /* void displayList() {
+            SQLiteHelper helper = new SQLiteHelper(this);
+            SQLiteDatabase db = helper.getReadableDatabase();
+            Cursor cursor = db.rawQuery("SELECT * FROM MYDRUG", null);
 
+            MonthAdapter adapter = new MonthAdapter();
+
+            while(cursor.moveToNext()){
+                //num 행은 가장 첫번째에 있으니 0번이 되고, name은 1번
+                adapter.addItemToList(cursor.getString(0),cursor.getString(1));
+            }
+        }
+
+        */
 
 }
- */
