+++ app/src/main/java/com/opensource13/pillsogood/MonthItem.java
@@ -1,0 +1,28 @@
+package com.opensource13.pillsogood;
+
+/**
+ * 일자 정보를 담기 위한 클래스 정의
+ */
+public class MonthItem {
+    private int dayValue;
+
+    public MonthItem() {
+
+    }
+
+    public MonthItem(int day) {
+
+        dayValue = day;
+    }
+
+    public int getDay() {
+
+        return dayValue;
+    }
+
+    public void setDay(int day) {
+
+        this.dayValue = day;
+    }
+
+}
